Refactoring into shared library for common operations
