+++ prov-api/helpers/workflows.py
@@ -1,12 +1,10 @@
 from ProvenaInterfaces.ProvenanceAPI import *
 from ProvenaInterfaces.ProvenanceModels import *
-from ProvenaInterfaces.SharedTypes import Status
 from helpers.entity_validators import *
 from helpers.registry_helpers import *
 from helpers.prov_helpers import model_run_to_graph
 from helpers.model_run_helpers import *
 from helpers.prov_connector import Neo4jGraphManager, GraphDiffApplier
-from services.encryption import EncryptionService
 from config import Config
