+++ registry-api/dependencies/dependencies.py
@@ -1,11 +1,12 @@
 from ProvenaInterfaces.AuthAPI import ENTITY_REGISTRY_COMPONENT, AccessLevel
+from ProvenaInterfaces.SharedTypes import UserInfo
 from KeycloakFastAPI.Dependencies import User, build_keycloak_auth, build_test_keycloak_auth, ProtectedRole
 from config import base_config, Config, get_settings
 from helpers.keycloak_helpers import setup_secret_cache
-from ProvenaInterfaces.SharedTypes import UserInfo
-from services.encryption import EncryptionService, KMSEncryptionService, KMSConfig
-from helpers.encryption_helpers import decrypt_user_dep
+from ProvenaSharedFunctionality.Services.encryption import EncryptionService, KMSEncryptionService, KMSConfig
+from ProvenaSharedFunctionality.Helpers.encryption_helpers import encrypt_user_info, decrypt_user_info
 from fastapi import Depends, Request, HTTPException
+from typing import Dict
 
 # Setup auth -> test mode means no sig enforcement on validation
 kc_auth = build_keycloak_auth(
@@ -114,7 +115,43 @@ async def get_user_context(
         )
 
     # try and decrypt it
-    decrypted = await decrypt_user_dep(cipher_text=header_value, encryption_service=encryption_service)
+    user_info = await decrypt_user_info(cipher_text=header_value, encryption_service=encryption_service)
 
     # NOTE this token is not valid - other details are decrypted from user info payload
-    return decrypted
+    return ProtectedRole(access_roles=user_info.roles, user=User(username=user_info.username, roles=user_info.roles, access_token=""))
+
+
+def get_user_context_header(user_cipher: str, config: Config) -> Dict[str, str]:
+    """
+
+    Returns a dictionary intended to be merged with other headers which includes
+    the user cipher at the configured user context header field.
+
+    Parameters
+    ----------
+    user_cipher : str
+        The encrypted user context
+    config : Config
+        The config
+
+    Returns
+    -------
+    Dict[str, str]
+        The headers
+    """
+    return {config.user_context_header: user_cipher}
+
+
+async def get_user_cipher(
+    encryption_service: EncryptionService = Depends(get_encryption_service),
+    # check user is defined through dependency
+    user: User = Depends(user_general_dependency)
+) -> str:
+    # use the encryption service to build the cipher
+    return await encrypt_user_info(
+        payload=UserInfo(
+            email=user.email,
+            roles=user.roles,
+            username=user.username
+        ),
+        encryption_service=encryption_service)
